legacy/requirements.txt
# Web server
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
aiohttp==3.11.15

# LLM integration
ollama==0.4.7

# HTML/Text processing
beautifulsoup4==4.13.3
html2text==2024.2.26
lxml==5.3.1

# Utilities
loguru==0.7.3
pydantic==2.11.1
python-dotenv==1.1.0
requests==2.32.3

trafilatura==2.0.0
semantic-text-splitter==0.25.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
mdit-plain==1.0.1
googlesearch-python==1.3.0
# torch==2.6.0
# transformers==4.50.3
torch==2.1.2
torchvision==0.16.2
transformers==4.41.0
--extra-index-url https://download.pytorch.org/whl/cu121
playwright==1.51.0
spacy==3.8.4
RapidFuzz==3.12.2
stanza==1.10.1

# Optional - search API integration
google-api-python-client==2.166.0

# numpy==2.2.4
numpy==1.26.4
# networkx==3.4.2
networkx==3.2.1
tqdm==4.67.1
scikit-learn==1.6.1
sentence-transformers==4.0.1
accelerate==0.26.0
optree==0.15.0